requirements.txt
flask
geopandas
Pillow
exif
